requirements.txt
alpaca-py
pandas
numpy
ta
requests
websocket-client
pytz
